+++ NEWS.md
@@ -7,6 +7,10 @@ NEWS
 - filtra_d usar funcao em C++ para calc distancias
 - Revisar Ingles e terminar de traduzir
 
+### percolator 0.4.9 (Release date: 2022-08-30)
+- Fixed s2 polygon problem in LT function by adding sf::sf_use_s2(FALSE)
+- Add dist_max filter in filtra_saag() function
+
 ### percolator 0.4.8 (Release date: 2022-08-18)
 - Create filtra_saag function for obtaining MCSs to SAAG Deep Convection Working 
   Group project.